Add provider package to allow for being imported
The provider package allows for the plugin to be imported by another
program, and have that program hand the plugin an existing net.Listener
for incoming traffic. In that scenario, the plugin no longer has to
worry about contructing the listener, or worry about endpoints.

The `goioc` package is used to register with a type registry on package
import (via the package init() method) so that the importing program can
instantiate the plugin/provider by name, rather than type, and it can
assert that the provider adheres to a specific interface.